アカウントページのインポートパスを修正し、コース作成APIにおいてインストラクターの取得方法をユーザー情報に基づいて更新しました。また、必須フィールドのチェックを追加し、エラーハンドリングを強化しました。不要なコンポーネントを削除しました。
