+++ src/app/api/courses/route.ts
@@ -1,11 +1,35 @@
+import { createClient } from "@/utils/supabase/server";
 import { prisma } from "@/lib/prisma";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function POST(request: NextRequest) {
-  const { title, description, price, videoUrl } = await request.json();
+  const supabase = await createClient();
+  const {
+    data: { user },
+    error: userError,
+  } = await supabase.auth.getUser();
+
+  if (userError) {
+    return NextResponse.json({ error: userError.message }, { status: 401 });
+  }
+
+  // フォームから送信されたデータ
+  const formData = await request.formData();
+  const title = formData.get("title");
+  const description = formData.get("description");
+  const price = formData.get("price");
+  const videoUrl = formData.get("videoUrl");
+
+  if (!title || !description || !price || !videoUrl) {
+    return NextResponse.json(
+      { error: "Missing required fields" },
+      { status: 400 }
+    );
+  }
+
   const instructor = await prisma.user.findUnique({
     where: {
-      email: "test@test.com",
+      email: user?.email,
     },
   });
 
@@ -15,17 +39,21 @@ export async function POST(request: NextRequest) {
       { status: 404 }
     );
   }
-  const instructorId = instructor.id;
 
   await prisma.course.create({
     data: {
-      title,
-      description,
-      videoUrl,
-      price,
+      title: title as string,
+      description: description as string,
+      videoUrl: videoUrl as string,
+      price: Number(price),
       instructor: {
-        connect: { id: instructorId },
+        connect: { id: instructor.id },
       },
     },
   });
+
+  return NextResponse.json(
+    { message: "Course created successfully" },
+    { status: 201 }
+  );
 }
